Fix issue with accessing collections.abc.Hashable
I assume this was broken by my earlier change to silence a Python warning.

It only breaks if you call yaml.safe_load() with a file handle instead
of a filename, such as:

with open("myfile.yaml", "r") as c:
    x = yaml.safe_load(c)

I tried to create a new test here in pyyaml's test suite to show this
failure, but could never get it to fail in that context.